Create Delete Journal Entries
Use this .js in a Roll20 game's mod section. 
It can be used to delete characters in a semi-bulk, semi-safe way. 
Use "!deleteCharacters a" (minus quotes) to delete all Character journal entries starting with the letter 'a'.
With "!sortCharacters a" or "!sortCharacters z" (minus quotes), the script will alphabetize or reverse-alphabetize the journal Characters.